Corrected Security Entry

+++ ServicesLayer/Services/Service.vb
@@ -172,7 +172,7 @@ Namespace Services
             Return dataModel
         End Function
 
-        Public Function GetUserSecurity(securityObjectIdNo As Int16, securityGroupIdNo As Int16) As ArrayList Implements IService.GetUserSecurity
+        Public Function GetUserSecurity(securityObjectIdNo As Int32, securityGroupIdNo As Int16) As ArrayList Implements IService.GetUserSecurity
             Return BaseDao.GetUserSecurity(securityObjectIdNo, securityGroupIdNo)
         End Function
